put more logic in useColumns
